Refactor and isolate a snapshot manager class
